refactor(services): extract GitHub I/O to community-api.ts (B-025)
First slice of the architect's "single biggest refactor" — split
`services/community-packages.ts` (which was doing four jobs in one
355-line file) into focused modules.

This commit moves the pure GitHub I/O concerns out:

- `loadCommunityPackagesFromGitHub()` — calls Contents API, fetches
  each .yml, parses, returns PackageItem[]
- `isAllowedDownloadUrl()` + `ALLOWED_DOWNLOAD_HOSTS` — the S-005
  host-allowlist defence against MITM'd Contents API responses
- `GitHubContentEntry` interface
- `convertSnippetsToObject()` — YAML-shape → flat snippet map
  (exported so the vault-backed `loadDynamicCommunityPackages` in
  community-packages.ts can still call it without copying the
  helper)

Also refactored a bit during the move:
- Extracted per-pack `fetchAndParsePack()` helper from the big for-
  loop body — makes the parallel-fetch refactor in the next commit
  (B-024) a much smaller diff.
- The GitHub Contents API URL is now a module-level constant.

`community-packages.ts` becomes a smaller facade — keeps
`loadDynamicCommunityPackages` (vault-backed loader), the cache
wrapper, the `loadAllCommunityPackages` router, and the dead
`createPackageIssue` (slated for deletion in a later commit).

Test imports updated to point at `community-api.ts` for the
relocated function. 350/350 tests still pass, coverage gates met
aggregate (lines 93.28 / branches 88.77 / functions 96.2).
