+++ src/services/community-packages.ts
@@ -1,40 +1,14 @@
 import * as YAML from "yaml";
 import { App, TFolder, TFile, requestUrl } from "obsidian";
 import type { PackageData } from "./package-types";
-import { normalizeTrigger } from "../engine/triggers";
+import { loadCommunityPackagesFromGitHub, convertSnippetsToObject } from "./community-api";
 
-/** Hostnames allowed for the second `requestUrl` fetch in
- *  `loadCommunityPackagesFromGitHub`. The Contents API returns a
- *  `download_url` that is *intended* to be a raw.githubusercontent.com
- *  URL — but the response is JSON over HTTPS to api.github.com, and a
- *  MITM (or a future API change) could return a different host. We
- *  treat this as untrusted and validate before fetching. See S-005.
- */
-const ALLOWED_DOWNLOAD_HOSTS = new Set<string>([
-  "raw.githubusercontent.com",
-]);
-
-function isAllowedDownloadUrl(url: unknown): boolean {
-  if (typeof url !== "string") return false;
-  try {
-    const parsed = new URL(url);
-    if (parsed.protocol !== "https:") return false;
-    return ALLOWED_DOWNLOAD_HOSTS.has(parsed.hostname);
-  } catch {
-    return false;
-  }
-}
-
-// No built-in packages - all packages come from GitHub API
+// GitHub I/O (loadCommunityPackagesFromGitHub, isAllowedDownloadUrl,
+// convertSnippetsToObject) moved to `community-api.ts` in 1.1.7
+// (B-025) per the architect's "single biggest refactor".
 
 interface GitHubIssueResponse {
   html_url: string;
-}
-
-interface GitHubContentEntry {
-  name: string;
-  download_url: string;
-  type?: string;
 }
 
 export interface PackageItem {
@@ -160,87 +134,6 @@ export async function createPackageIssue(_app: App, packageData: PackageData, us
 }
 
 /**
- * Loads community packages from GitHub API
- */
-export async function loadCommunityPackagesFromGitHub(): Promise<PackageItem[]> {
-  try {
-    const response = await requestUrl({
-      url: 'https://api.github.com/repos/Dimagious/snipsidian-community/contents/community-packages/approved'
-    });
-    
-    if (response.status !== 200) {
-      if (response.status === 404) {
-        return [];
-      }
-      throw new Error(`GitHub API error: ${response.status}`);
-    }
-
-    const files = JSON.parse(response.text) as GitHubContentEntry[];
-
-    // Handle case where directory exists but is empty
-    if (!Array.isArray(files) || files.length === 0) {
-      return [];
-    }
-
-    const packages: PackageItem[] = [];
-
-    for (const file of files) {
-      if (file.name.endsWith('.yml') || file.name.endsWith('.yaml')) {
-        try {
-          // Hostname allowlist: refuse to follow any URL outside the
-          // expected raw.githubusercontent.com host. Defends against a
-          // spoofed `api.github.com` response that points to an attacker
-          // host. See security S-005.
-          if (!isAllowedDownloadUrl(file.download_url)) {
-            console.error(`[snipsy] refusing to fetch ${file.name}: download_url is not on allowlist`, file.download_url);
-            continue;
-          }
-          const contentResponse = await requestUrl({
-            url: file.download_url
-          });
-          const content = contentResponse.text;
-          const packageData = YAML.parse(content) as PackageData;
-
-          if (packageData && packageData.name && packageData.snippets) {
-            const snippets = convertSnippetsToObject(packageData.snippets);
-            
-            // Only add package if it has snippets
-            if (Object.keys(snippets).length > 0) {
-              const tags = Array.isArray(packageData.tags) 
-                ? packageData.tags.filter((t): t is string => typeof t === 'string')
-                : typeof packageData.tags === 'string'
-                  ? [packageData.tags]
-                  : [];
-              const packageItem: PackageItem = {
-                id: file.name.replace(/\.(yml|yaml)$/, ''),
-                label: packageData.name,
-                description: packageData.description,
-                author: packageData.author,
-                version: packageData.version,
-                downloads: 0, // Will be updated when packages are actually downloaded
-                tags: tags,
-                verified: true,
-                rating: 0, // Will be updated based on actual user ratings
-                snippets: snippets
-              };
-
-              packages.push(packageItem);
-            }
-          }
-        } catch (error) {
-          console.error(`Failed to load package ${file.name}:`, error);
-        }
-      }
-    }
-
-    return packages;
-  } catch (error) {
-    console.error('Failed to load community packages from GitHub:', error);
-    return [];
-  }
-}
-
-/**
  * Loads community packages with caching
  */
 interface PluginWithApp {
@@ -307,46 +200,6 @@ export async function loadAllCommunityPackages(app: App, plugin?: PluginWithApp)
     return [];
   }
 }
-
-/**
- * Converts YAML snippets to object format
- * Handles both array format and object format
- */
-function convertSnippetsToObject(snippets: PackageData['snippets']): { [trigger: string]: string } {
-  const snippetsObj: { [trigger: string]: string } = {};
-
-  // B-117: triggers are normalised through `normalizeTrigger` so Espanso-style
-  // keys (`:smile:`, `:fire`, etc.) become reachable. Snipsy's engine treats `:`
-  // as a separator, so any leading / trailing colons in the stored key would
-  // make the trigger unmatchable via keystroke expansion. The Espanso *importer*
-  // already does this (`src/packages/espanso.ts`); the install path for
-  // community packages historically did not, which is why `basic-emojis.yml`
-  // was effectively dead code after install.
-  const recordKey = (rawTrigger: string, replacement: string) => {
-    const key = normalizeTrigger(rawTrigger);
-    if (!key) return; // pure-colon / empty triggers can't be reached anyway
-    snippetsObj[key] = replacement;
-  };
-
-  if (Array.isArray(snippets)) {
-    // Array format: [{ trigger: "...", replace: "..." }]
-    for (const snippet of snippets) {
-      if (snippet.trigger && snippet.replace) {
-        recordKey(snippet.trigger, snippet.replace);
-      }
-    }
-  } else if (snippets && typeof snippets === 'object') {
-    // Object format: { "trigger": "replacement", ... }
-    for (const [trigger, replacement] of Object.entries(snippets)) {
-      if (typeof replacement === 'string') {
-        recordKey(trigger, replacement);
-      }
-    }
-  }
-
-  return snippetsObj;
-}
-
 
 /* `processPackageSubmission` removed in 1.0.8 — was wired only to the
  * dead `SubmitPackageModal` (also removed). Active submission flow
